Rename package to com.koriit.positioner.android
